src/requirements.txt
Flask
flask_cors
pyjwt
python-dotenv
requests
flask-talisman
Crypto
py-cid
py-multibase
pymongo